% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt data}
\usage{
decrypt(x, password)
}
\arguments{
\item{x}{any data object.}

\item{password}{a string specifying the passphrase to en- and decrypt
the data.}
}
\value{
encrypted data
}
\description{
\code{decrypt} is a simplified wrapper for the decryption of R data
objects using the \code{sodium} package
}
\seealso{
\code{\link{encrypt}}
}

